+++ src/main/java/com/clinica/service/iService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface iService<T> {
     T guardar(T t);
-    T buscar(Integer id);
-    void eliminar(Integer id);
+    T buscar(Long id);
+    void eliminar(Long id);
     void actualizar(T t);
     T buscarGenerico(String parametro);
     List<T> buscarTodos();
